enh: added SparseGeometry.add, changed `align` to `offset`
This enables direct addition of two sparse matrices
without having to worry about the couplings.
I.e. both sparse matrices will be added as is.
Note that if the lattice vectors are not commensurate
then supercell couplings may not be transferred
as one might expect.
Currently there is no warning for this, but possibly
we could warn if couplings are being discarded.

The `align` keyword argument in SparseGeometry.append/prepend
has now changed to `offset` (also for Geometry.append/prepend)
This will enable a consistent argument that makes sense
for its purpose and how it may be used.
For Geometry it may either be a string (none or min) or a
vector.
The `offset` argument has also been added to Geometry.add
method.

Added tests for SparseGeometry.add (with and without axis
specification).

+++ sisl/io/siesta/binaries.py
@@ -1143,7 +1143,7 @@ class _gfSileSiesta(SileBinSiesta):
         """
         no = len(H)
         if S is None:
-            S = np.eye(no, dtype=np.complex128)
+            S = np.eye(no, dtype=H.dtype)
         self._step_counter('write_hamiltonian', HS=True, read=True)
         _siesta.write_gf_hs(self._iu, self._ik, self._E[self._iE],
                             H.astype(np.complex128, 'C', copy=False).T,
